+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Reorganized settings page layout for better user experience.
 
+### Fixed
+
+- Fixed the issue where some articles failed to open when filter is applied. (#304)
+
 ## [1.3.8] - 2025-12-24
 
 ### Added